+++ crates/generator-shared/src/select_include/mod.rs
@@ -1,0 +1,5 @@
+mod selectable_fields;
+mod variant;
+
+pub use selectable_fields::*;
+pub use variant::*;
